changes in CSS and add api call to backend
