+++ package.json
@@ -12,6 +12,7 @@
     "@types/node": "18.15.11",
     "@types/react": "18.0.32",
     "@types/react-dom": "18.0.11",
+    "axios": "^1.3.5",
     "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
