+++ src/styles/home-page.scss
@@ -3,7 +3,7 @@
     padding: 25px;
     min-height: 100vh;
     background: linear-gradient(#371a54 , #2d005a, #262b5e);
-    
+
     .container{
         display: flex;
         justify-content: space-between;
@@ -15,7 +15,7 @@
         margin: 0 auto;
     }
 
-    .header{   
+    .header{
         display: flex;
         justify-content: space-between;
         align-items: center;
@@ -27,7 +27,7 @@
             justify-content: space-between;
             border: 15px;
             border-radius: 25px;
-    
+
             .btn-link{
                 font-weight: 600;
                 text-decoration: none;
@@ -39,7 +39,7 @@
 
                 &:hover{
                     background-color: #ffa02ce5;
-        
+
                 }
             }
         }
@@ -58,12 +58,11 @@
                 position: absolute;
                 top: 20px;
                 right: 0;
-    
+
                 display: flex;
                 width: 35px;
                 background-color: unset;
                 border: 0;
-
             }
 
             .hamburguer-btn, .hamburguer-close-btn{
@@ -73,7 +72,7 @@
                 flex-direction: column;
                 gap: 10px;
                 opacity: 1;
-                
+
                 .line{
                     transition: all 0.2s ease-in-out;
                     width: 100%;
@@ -162,12 +161,13 @@
                 display: none;
             }
         }
-    }   
+    }
 
     .section-content{
         width: 100%;
         padding-top: 150px;
-    
+        padding-right: 300px;
+
         .content{
             width: 100%;
             max-width: 600px;
@@ -183,7 +183,7 @@
                 width: 100%;
             }
         }
-    
+
         .box-title {
             font-weight: 800;
             color: white;
@@ -196,7 +196,7 @@
                 }
             }
         }
-    
+
         .box-description{
             text-align: center;
             margin-top: 15px;
@@ -216,7 +216,7 @@
             &:hover{
                 background-color: #ffa02ce5;
             }
-        
+
             .log-link {
                 padding: 8px 0;
                 font-weight: 600;
@@ -227,13 +227,40 @@
             }
         }
     }
-       
-    @media(max-width: 768px) {        
+
+    @media(max-width: 1500px) {
         .section-content{
             padding-top: 20px;
+            padding-right: 80px;
+            margin-left: 80px;
+        }
+    }
+
+    @media(max-width: 1000px) {
+        .section-content{
+            padding-top: 20px;
+            padding-right: 75px;
+            margin-left: 30px;
 
             .container{
                 flex-direction: column;
+            }
+        }
+    }
+
+    @media(max-width: 400px) {
+        .section-content{
+            padding-top: 20px;
+            padding-right: 75px;
+            margin-left: 30px;
+
+            .container{
+                flex-direction: column;
+            }
+            .box-title {
+                h1 {
+                    font-size: 2.75rem;
+                }
             }
         }
     }
