+++ src/styles/login-page.scss
@@ -45,7 +45,7 @@
                 font-weight: bold;
                 font-size: 18px;
             }
-            
+
             input{
                 margin-top: 5px;
                 width: 100%;
@@ -56,7 +56,7 @@
                 background-color: white;
                 border-radius: 5px;
                 box-shadow: 0 0 2px black;
-                
+
                 &:focus-within{
                     box-shadow: 0 0 5px black;
                 }
@@ -71,7 +71,7 @@
                 position: relative;
             }
 
-            
+
             .wrapper{
                 display: flex;
                 flex-direction: column;
@@ -79,15 +79,15 @@
                 .btn-submit{
                     padding: 8px;
                     background-color: #673897;
-    
+
                     color: white;
                     font-weight: bold;
                     font-size: 0.875rem;
-    
+
                     border: 0;
                     border-radius: 5px;
                     cursor: pointer;
-    
+
                     &:hover{
                         background-color: #8154ad;
                     }
