+++ src/styles/signup-page.scss
@@ -13,7 +13,7 @@
         display: flex;
         justify-content: center;
         align-items: center;
-        
+
         .box-title{
             display: flex;
             flex-direction: column;
@@ -93,14 +93,14 @@
         width: 30%;
         background-color: #371A54;
         box-shadow: 0 0 10px black;
-        min-height:  100%;   
+        min-height:  100%;
 
         .box-description{
             display: flex;
             flex-direction: column;
             justify-content: center;
             align-items: center;
-            
+
 
            img{
             cursor: pointer;
