+++ src/types/login.ts
@@ -1,0 +1,4 @@
+export interface UserAuthInput {
+    email: string;
+    password: string;
+}
