+++ src/utils/base64-encode.ts
@@ -1,0 +1,3 @@
+export const  base64Encode = (str: string): string => {
+    return btoa(btoa(str))
+}
