+++ src/utils/redirect.ts
@@ -1,0 +1,5 @@
+import { NextRouter} from "next/router"
+
+export const redirect = (router: NextRouter,to: string) => {
+    router.push(to)
+}
